requirements.txt
strands-agents>=1.0.0
boto3>=1.34.0
requests>=2.31.0
jsonschema>=4.21.0
